Adding memory of requests and auto-recover.
Now users don't have to track the connection state, or protect against
making requests in the wrong state.  They just make a request and the
library will either start browsing, or queue the request for when the
client does connect.  For disconnects, we now automatically save the
browser's info before halting the browsers, and then we restore it on
reconnect.